requirements.txt
Django==1.11.4
Pillow==4.3.0
django-debug-toolbar